backend/requirements.txt
flask
python-dotenv
pymongo
flask-bcrypt
google-genai